Multi-threshold support
This is a non-backwards compatible change.

The `thresholds` setting is now a mapping of labels to value to
background colors:

    thresholds={'build': { 'SUCCESS': '#2c2', 'FAILURE': '#c22' },
               {'test': { 'SUCCESS': '#22c', 'FAILURE': '#860' }}

This allows one instance (or config) to generate badges for all label
types.

The `value_backgrounds` acts as a fallback if the label + value is not
found in the `thresholds` data structure:

    value_backgrounds={ 'SUCCESS': 'green', 'FAILURE': 'red' }

Closes #1 Multi-threshold support

Signed-off-by: Gunnar Strand <Gurra.Strand@gmail.com>